pyproject.toml
[project]
name = "meteo-haiti-v2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.1",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "pydeck>=0.9.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
]
